video_gen/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
openai>=1.0.0
aiohttp>=3.9.0
moviepy>=1.0.3
Pillow>=10.0.0
edge-tts>=6.1.0
python-dotenv>=1.0.0
requests>=2.31.0
